# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowAccess API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Plan {
  id: number;
  name: string;
  priceUsd: string;
  creditsPerMonth: number;
  /** @nullable */
  description?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface UserProfile {
  id: number;
  clerkUserId: string;
  email: string;
  isAdmin: boolean;
  /** @nullable */
  planId?: number | null;
  /** @nullable */
  planName?: string | null;
  creditsTotal: number;
  creditsUsed: number;
  creditsRemaining: number;
  /** @nullable */
  subscriptionStartedAt?: string | null;
  createdAt: string;
}

export interface ApiTokenResponse {
  token: string;
}

export interface InjectResponse {
  cookieData: string;
  creditsRemaining: number;
}

export interface AdminStats {
  totalSessions: number;
  activeSessions: number;
  totalUsers: number;
  totalUsage: number;
}

export interface AdminSession {
  id: number;
  label: string;
  cookieData: string;
  isActive: boolean;
  /** @nullable */
  lastUsedAt?: string | null;
  usageCount: number;
  /** @nullable */
  syncKey?: string | null;
  /** @nullable */
  cookieUpdatedAt?: string | null;
  createdAt: string;
}

export interface CreateSessionBody {
  label: string;
  cookieData: string;
  isActive?: boolean;
}

export interface UpdateSessionBody {
  label?: string;
  cookieData?: string;
  isActive?: boolean;
}

export interface AdminUser {
  id: number;
  clerkUserId: string;
  email: string;
  isAdmin: boolean;
  /** @nullable */
  planId?: number | null;
  /** @nullable */
  planName?: string | null;
  creditsTotal: number;
  creditsUsed: number;
  /** @nullable */
  subscriptionStartedAt?: string | null;
  createdAt: string;
}

export interface CreateUserBody {
  email: string;
  password: string;
  /** @nullable */
  planId?: number | null;
  creditsTotal?: number;
  isAdmin?: boolean;
}

export interface UpdateUserBody {
  /** @nullable */
  planId?: number | null;
  creditsTotal?: number;
  creditsUsed?: number;
  isAdmin?: boolean;
}

export interface AdminUsageLog {
  id: number;
  userId: number;
  userEmail: string;
  /** @nullable */
  sessionId?: number | null;
  action: string;
  creditsUsed: number;
  createdAt: string;
}

export interface UsageLogEntry {
  id: number;
  /** @nullable */
  sessionId?: number | null;
  action: string;
  creditsUsed: number;
  createdAt: string;
}

export type GenerateSessionSyncKey200 = {
  syncKey: string;
};