# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowAccess API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSession {
  id: number;
  label: string;
  cookieData: string;
  isActive: boolean;
  /** @nullable */
  lastUsedAt?: Date | null;
  usageCount: number;
  /** @nullable */
  syncKey?: string | null;
  /** @nullable */
  cookieUpdatedAt?: Date | null;
  createdAt: Date;
}